# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HKUST-NFT Server API",
        "title": "HKUST-NFT",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/collections/{collection-id}": {
            "get": {
                "description": "single collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collection"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "collection id",
                        "name": "collection-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET/api/v1/collections/hahahah",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/collections/{collection-id}/items": {
            "get": {
                "description": "all items in collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "collection"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "collection id",
                        "name": "collection-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET/api/v1/collections/hahahah/items",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/{id}/change-profile": {
            "post": {
                "description": "change profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user Name",
                        "name": "userName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "POST/api/v1/Anna/change-profile",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{id}/collected": {
            "get": {
                "description": "collected",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET/api/v1/Anna/collected",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{id}/create-collection": {
            "post": {
                "description": "create-collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "collection Name",
                        "name": "collection-Name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "POST/api/v1/Anna/create-collection",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{id}/create-item": {
            "post": {
                "description": "create-item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "item Name",
                        "name": "item-Name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "POST/api/v1/Anna/create-item",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{id}/creation": {
            "get": {
                "description": "creation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET/api/v1/Anna/creation",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{id}/edit-collection": {
            "post": {
                "description": "edit-collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "collection Name",
                        "name": "collection-Name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "POST/api/v1/Anna/edit-collection",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{id}/edit-item": {
            "post": {
                "description": "edit-item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "item Name",
                        "name": "item-Name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "POST/api/v1/Anna/edit-item",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{id}/favorites": {
            "get": {
                "description": "favorites",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GET/api/v1/Anna/favorites",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
...
}